# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: StarShield API
  description: |
    星盾（StarShield）海量游戏玩家发言舆情与违规智能监控中台 API 契约文档。
    约束：所有接口变更必须先更新本文件，再修改代码。
  version: 1.1.0

servers:
  - url: http://localhost:8080
    description: 本地开发环境

tags:
  - name: Chat
    description: 玩家发言接收接口
  - name: AdminModeration
    description: 管理后台审核工作流
  - name: ControlPanel
    description: 动态规则与 Prompt 热管理
  - name: Archive
    description: 海量检索中台
  - name: Dashboard
    description: 大屏实时指标

paths:
  /api/chat/upload:
    post:
      tags: [Chat]
      summary: 上传玩家发言（高并发入口）
      operationId: uploadChatMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageUploadRequest'
      responses:
        '200':
          description: 接收成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
        '429':
          description: 限流
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
        '500':
          description: 接收失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'

  /api/admin/moderation/idempotency-key:
    get:
      tags: [AdminModeration]
      summary: 生成幂等键
      operationId: getModerationIdempotencyKey
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultMapString'

  /api/admin/moderation/pending:
    get:
      tags: [AdminModeration]
      summary: 查询待复核列表
      operationId: getPendingModerationRecords
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultChatMessageList'

  /api/admin/moderation/{id}/audit-logs:
    get:
      tags: [AdminModeration]
      summary: 查询某条记录操作审计日志
      operationId: getModerationAuditLogs
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultModerationAuditList'

  /api/admin/moderation/{id}/confirm-ban:
    post:
      tags: [AdminModeration]
      summary: 人工确认封禁
      operationId: confirmBan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
        '404':
          description: 记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
        '409':
          description: 幂等键无效或重复
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'

  /api/admin/moderation/{id}/release:
    post:
      tags: [AdminModeration]
      summary: 人工解除封禁
      operationId: releaseBan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
        - in: header
          name: X-Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
        '404':
          description: 记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
        '409':
          description: 幂等键无效或重复
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'

  /api/control/rules/sensitive-words:
    get:
      tags: [ControlPanel]
      summary: 获取敏感词列表
      operationId: getSensitiveWords
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultStringList'
    put:
      tags: [ControlPanel]
      summary: 全量替换敏感词列表
      operationId: replaceSensitiveWords
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensitiveWordsRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'

  /api/control/prompt:
    get:
      tags: [ControlPanel]
      summary: 获取当前 Prompt
      operationId: getPrompt
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultMapString'
    put:
      tags: [ControlPanel]
      summary: 更新当前 Prompt
      operationId: updatePrompt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromptRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'

  /api/archive/search:
    get:
      tags: [Archive]
      summary: 组合条件检索
      operationId: archiveSearch
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: playerId
          schema:
            type: string
        - in: query
          name: decision
          schema:
            type: string
            enum: [PASS, REVIEW, BLOCK]
        - in: query
          name: labels
          schema:
            type: string
        - in: query
          name: startTime
          schema:
            type: string
            format: date-time
        - in: query
          name: endTime
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 200
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultChatMessageList'

  /api/archive/search/highlight:
    get:
      tags: [Archive]
      summary: 组合条件检索（含 ES 高亮片段）
      operationId: archiveSearchHighlight
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: playerId
          schema:
            type: string
        - in: query
          name: decision
          schema:
            type: string
            enum: [PASS, REVIEW, BLOCK]
        - in: query
          name: labels
          schema:
            type: string
        - in: query
          name: startTime
          schema:
            type: string
            format: date-time
        - in: query
          name: endTime
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 200
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultArchiveSearchHitList'

  /api/archive/analysis:
    get:
      tags: [Archive]
      summary: 归档聚合分析
      operationId: archiveAnalysis
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
        - in: query
          name: playerId
          schema:
            type: string
        - in: query
          name: decision
          schema:
            type: string
            enum: [PASS, REVIEW, BLOCK]
        - in: query
          name: labels
          schema:
            type: string
        - in: query
          name: startTime
          schema:
            type: string
            format: date-time
        - in: query
          name: endTime
          schema:
            type: string
            format: date-time
        - in: query
          name: topHitLimit
          schema:
            type: integer
            default: 5
            maximum: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultArchiveAnalysis'

  /api/archive/reindex:
    post:
      tags: [Archive]
      summary: 从 MySQL 回填归档数据到 ES
      operationId: archiveReindex
      parameters:
        - in: query
          name: batchSize
          schema:
            type: integer
            default: 500
            maximum: 1000
        - in: query
          name: maxRows
          schema:
            type: integer
            default: 10000
            maximum: 1000000
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultMapObject'

  /api/dashboard/metrics:
    get:
      tags: [Dashboard]
      summary: 获取大屏实时指标
      operationId: getDashboardMetrics
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultMapObject'

components:
  schemas:
    ResultVoid:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          nullable: true

    ResultStringList:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            type: string

    ResultMapString:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          additionalProperties:
            type: string

    ResultMapObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          additionalProperties: true

    ResultChatMessageList:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageLog'

    ResultArchiveSearchHitList:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ArchiveSearchHit'

    ArchiveSearchHit:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/ChatMessageLog'
        highlights:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        highlightContent:
          type: string

    ResultArchiveAnalysis:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/ArchiveAnalysis'

    ArchiveAnalysis:
      type: object
      properties:
        platformDistribution:
          type: object
          additionalProperties:
            type: integer
            format: int64
        timeTrend:
          type: array
          items:
            $ref: '#/components/schemas/ArchiveTimeBucket'
        topHits:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageLog'

    ArchiveTimeBucket:
      type: object
      properties:
        time:
          type: string
        count:
          type: integer
          format: int64

    ResultModerationAuditList:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ModerationAuditLog'

    ChatMessageUploadRequest:
      type: object
      required: [playerId, content, platform]
      properties:
        playerId:
          type: string
          maxLength: 64
        content:
          type: string
          maxLength: 10000
        platform:
          type: string
          enum: [GAME_INNER, BILIBILI, BILIBILI_LIVE, WEIBO, DOUYIN, OTHER]
        status:
          type: integer
          enum: [0, 1, 2]
          default: 0

    ChatMessageLog:
      type: object
      properties:
        id:
          type: integer
          format: int64
        playerId:
          type: string
        content:
          type: string
        platform:
          type: string
        status:
          type: integer
        decision:
          type: string
          enum: [PASS, REVIEW, BLOCK]
        riskScore:
          type: integer
        labels:
          type: string
        hitWords:
          type: string
        aiAnalysisResult:
          type: string
          nullable: true
        createTime:
          type: string
          format: date-time

    ModerationAuditLog:
      type: object
      properties:
        id:
          type: integer
          format: int64
        messageId:
          type: integer
          format: int64
        operator:
          type: string
        action:
          type: string
          enum: [CONFIRM_BAN, RELEASE]
        beforeDecision:
          type: string
          nullable: true
        afterDecision:
          type: string
          nullable: true
        beforeRiskScore:
          type: integer
          nullable: true
        afterRiskScore:
          type: integer
          nullable: true
        createTime:
          type: string
          format: date-time

    OperatorRequest:
      type: object
      properties:
        operator:
          type: string

    SensitiveWordsRequest:
      type: object
      properties:
        words:
          type: array
          items:
            type: string

    PromptRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string

x-contract-rules:
  immutable_fields:
    - playerId
    - content
    - platform
    - status
    - aiAnalysisResult
    - createTime
  note: |
    已约定字段名为契约字段，禁止随意重命名（例如 playerId 不能改为 player_id）。